chore: checkpoint dungeon fixes before class restoration

+++ src/server/game/Events/GameEventMgr.cpp
@@ -390,7 +390,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 ObjectGuid::LowType guid = fields[0].GetUInt64();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
